requirements.txt
streamlit
langchain-community
langchain-groq
langchain-huggingface
faiss-cpu
python-dotenv
pypdf
sentence-transformers
